Update calls to refactored core API

+++ src/main/java/tuwien/auto/calimero/device/ProcessCommunicationResponder.java
@@ -158,13 +158,13 @@ public class ProcessCommunicationResponder implements ProcessCommunicationBase
 	}
 
 	@Override
-	public void write(final GroupAddress dst, final float value, final boolean use4ByteFloat)
+	public void write(final GroupAddress dst, final double value, final boolean use4ByteFloat)
 		throws KNXTimeoutException, KNXFormatException, KNXLinkClosedException
 	{
 		if (use4ByteFloat) {
 			final DPTXlator4ByteFloat t = new DPTXlator4ByteFloat(
 					DPTXlator4ByteFloat.DPT_TEMPERATURE_DIFFERENCE);
-			t.setValue(value);
+			t.setValue((float) value);
 			write(dst, priority, t);
 		}
 		else {
